Put the structure of the project in order
