+++ src/Layout/index.jsx
@@ -1,0 +1,13 @@
+import Header from "../components/Header";
+import { Outlet } from "react-router-dom";
+
+const Layout = () => {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+};
+
+export { Layout };
