--- src/components/pages/Movie.jsx
@@ -1,14 +1,0 @@
-import { useParams } from "react-router-dom";
-
-
-export const Movie = () => {
-
-    //console.log(useParams());
-    const {id} = (useParams());
-
-    return(
-        <div>
-            <p>movie id = {id}</p> 
-        </div>
-    );
-};
